Decouple scaling execution from parallelism computation
